CORTX Provisioner Framework Skeleton (Source Dir Structure)
