+++ conf/cluster.yaml.sample
@@ -17,12 +17,14 @@ cluster:
             - message_bus
       storage:
         - name: cvg-01
+          type: ios
           metadata_devices:
             - /dev/sda
           data_devices:
             - /dev/sdb
             - /dev/sdc
         - name: cvg-02
+          type: ios
           metadata_devices:
             - /dev/sdx
           data_devices:
